+++ doc/changes.rst
@@ -6,6 +6,8 @@ Change Log
 ------------------
 
 * Fixed a minor variable name bug.
+* Ignore warnings produced by processing MANIFEST.in.
+* Fixed return value in cross_install.
 
 1.1.1 (2015-11-18)
 ------------------